Create a single definition of the docstring sections
parse_docstring_sections() divides the docstring into the main
non-overlapping sections that are to be further analysed. These are
text before the usage, the usage header ("usage:" or similar), the usage
body (" proc [options]") and text following the usage.

The intention of this partitioning is to facilitate restoring the option
description parsing behaviour of the last stable docopt release (0.6.2),
while retaining the current improved behaviour of not requiring a blank
line after the usage: section; and also removing a small quirk in the
0.6.2 parser, which is that option-defaults can occur before the usage:
section. (But this partitioning provides access to the text before the
usage section, so this behaviour could be retained if desired.)