Restore compatibility with docopt 0.6.2 docstrings
This commit uses parse_docstring_sections() and parse_options() to
parse docstrings accepted by docopt 0.6.2, while retaining docopt-ng's
improvements to supported syntax.

Currently, docopt-ng parses option-defaults using a strategy that was in
docopt's master branch, but considered unstable by the author, and was
not released in docopt. It looks for option descriptions in an
"options:" section, which is ended on the first blank line. This has
the side-effect that options defined in a man-page style — with blank
lines in-between — are not found. Neither are options outside an
options: section (docopt allows options to follow the usage with no
section heading).

parse_docstring_sections() is used to separate the usage section from
the rest of the docstring. The text before the usage is ignored. The
usage body (without its header) is parsed for the argument pattern and
the usage header with its body is used to print the usage summary help.
The text following the usage is parsed for options descriptions, using
parse_options(), which supports option the description syntax of both
docopt and the current docopt-ng.

Note that docopt 0.6.2 recognises option descriptions in the text
prior to the usage section, but this change does not, as it seems like
an unintended side-effect of the previous parser's implementation, and
seems unlikely to be used in practice.

The testcases have two cases added for docopt 0.6.2 compatibility.

This fixes https://github.com/jazzband/docopt-ng/issues/33