Implement backwards compatible option parsing
parse_options() parses option descriptions in a way that is compatible
with docopt 0.6.2, while also compatible with docopt-ng's current
behaviour.

The differences are:

  - docopt-ng requires option descriptions to be in an "options:"
    section, docopt allows them to be anywhere outside the "usage:"
    section.
  - docopt-ng requires options descriptions have leading whitespace,
    docopt allows them to start at column 0.
  - docopt-ng allows options descriptions to begin on the same line as
    a section heading, docopt does not. e.g. `options: --foo` is OK with
    docopt-ng.

parse_options() parses options following either docopt or docopt-ng's
behaviour. Although it expects the `docstring` argument to be a
sub-section of the overall docstring, so the caller is in control of
where in the docstring options are parsed from.