fix: handle options with descriptions on new line
This fixes a pre-existing bug in docopt-ng that wasn't previously
triggered by a test. Options in the options: section without a
description on the same line and with content on the following line
(e.g. a line-wrapped description) would be parsed as if the content on
the following line was part of the option name & arguments.

Option.parse() now ends the option name & arguments section on the end
of the line, not just on two consecutive spaces.