test: add examples of options not being parsed
This adds another case for test_parse_options demonstrating how
parse_options() parses docs which contain both options and text that
mentions options that should not be interpreted as options.

This includes an example (the '-b' option) where we do currently parse
an option mentioned in prose as a real option. In principle we could fix
this, but it's here for now to document current behaviour. And it's not
hard to work around when writing a usage doc - just need to adjust where
a line is wrapped.