dev: swap black to ruff, drop pre-commit
now that ruff has formatting capabilities, we can drop one more dependency.

Also, the version of black and ruff in the pre-commit config was
out of sync from the version pinned in pdm.lock. That is bad.
So just drop pre-commit, and run stuff manually, there
are only a few commands. But I can't remove it entirely,
since pre-commit.ci is enabled on all jazzband repos
and I can't turn it off. So there needs to be some content in there, so I added some of the generic builtin ones.

I had to adjust a few of the example docstrings
so that ruff wouldn't autoformat away the needed leading indentation.

I was lazy and just added a bunch of `# noqa: E721` for the
`type(match.value) == str` checks that should be isinstance()
checks, because when I tried isinstance, some tests failed.
The current method isn't that egregious.